+++ app/api/delete-cookie/route.ts
@@ -3,6 +3,7 @@ import { NextRequest, NextResponse } from "next/server"
 
 export async function POST(request: NextRequest) {
   const cookieName = await request.text()
-  cookies().delete(cookieName)
+  const cookieStore = await cookies()
+  cookieStore.delete(cookieName)
   return new NextResponse(null, { status: 204 })
 }
